requirements.txt
annotated-types==0.7.0
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
google-auth==2.47.0
google-genai==1.60.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.15.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pypdfform==4.3.1
python-dotenv==1.1.0
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==15.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
boto3>=1.42.0
